+++ CHANGES.md
@@ -12,12 +12,12 @@ Fast and flexible PHP library rendering plain ASCII tables.
   * Added option to separately control table column title align.
   * Custom column max width is now correctly handled.
   * Fixed width columns are now truncated automatically with ellipsis.
-  * Improved code creating row cells from plain string arrays. 
+  * Improved code creating row cells from plain string arrays.
   * Added more tests.
 
 * v1.1.0 (2022-10-17)
   * Corrected `\ArrayAccess` implementation.
   * Added more tests.
-
+  
 * v1.0.0 (2022-10-17)
   * Initial public release.